% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boottestgscm.r
\name{averageR2w}
\alias{averageR2w}
\title{For use in boottestgscm.}
\usage{
averageR2w(dat, B, indicatorx, indicatory, loadingx = FALSE,
  loadingy = FALSE, member)
}
\arguments{
\item{dat}{(n,p)-matrix, the values of the manifest variables.
The columns must be arranged in that way that the components of refl are (absolutely) increasing.}

\item{B}{(q,q) lower triangular matrix describing the interrelations of the latent variables:
b_ij = 1 regression coefficient of eta_j in the regression relation in which eta_i is the depend variable 
b_ij = 0 if eta_i does not depend on eta_j in a direct way  (b_ii = 0 !)}

\item{indicatorx}{vector describing with which exogenous composite the X-variables are connected}

\item{indicatory}{vector describing with which endogenous composite the Y-variables are connected}

\item{loadingx}{logical TRUE when there are loadings for the X-variables in the model}

\item{loadingy}{logical TRUE when there are loadings for the Y-variables in the model}

\item{member}{vector of length n, indicating the cluster the observation belongs to}
}
\value{
r scalar, 'global' r2 coefficiet of determination
}
\description{
\code{averageR2w} computes the weighted average of average of coefficients of determination  for the structural parts 
of a segmented GSC model
}
