% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValeMorelli.r
\name{FlDeriv}
\alias{FlDeriv}
\title{\code{FlDeriv}compute the Jacobian of the Fleishman transform  for a given set of coefficients b,c,d}
\usage{
FlDeriv(coef)
}
\arguments{
\item{coef}{vector with the coefficents for the Fleishman transform}
}
\value{
J  (3,3) Jacobian matrix of partial derivatives
}
\description{
\code{FlDeriv}compute the Jacobian of the Fleishman transform  for a given set of coefficients b,c,d
}
\examples{
coef <- c( 0.90475830, 0.14721082, 0.02386092)  
J <- FlDeriv( coef )

}
