% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValeMorelli.r
\name{Fleishman}
\alias{Fleishman}
\title{\code{Fleishman} computes the variance, skewness and kurtosis for a given set of of coefficients b,c,d for the Fleishman transform}
\usage{
Fleishman(coef)
}
\arguments{
\item{coef}{vector with the coefficents}
}
\value{
out vector with coefficients Var,Skew,Kurt
}
\description{
\code{Fleishman} computes the variance, skewness and kurtosis for a given set of of coefficients b,c,d for the Fleishman transform
}
\examples{
coef <- c( 0.90475830, 0.14721082, 0.02386092) 
out <- Fleishman( coef )

}
