% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data6.r
\docType{data}
\name{poloecfree}
\alias{poloecfree}
\title{Political and economical freedom.}
\format{A data frame with 14 variables and 91 cases:
\describe{
  \item{no}{number}                                                                            
  \item{country}{country}                                                                            
  \item{CP75}{Competition of Parties 1975} 
  \item{PR75}{Politial Rights 1975}
  \item{CL75}{Civil Liberties  1975}   
  \item{AoP75}{Amount of Privatisation 1975}
  \item{FFE75}{Freedom of Foreign Exchange 1975}
  \item{FCM75}{Freedom of Capital Movements 1975}
  \item{CP95}{Competition of Parties 1995} 
  \item{PR95}{Politial Rights 1995} 
  \item{CL95}{Civil Liberties 1995}  
  \item{AoP95}{Amount of Privatisation 1995}      
  \item{FFE95}{Freedom of Foreign Exchange 1995}  
  \item{FCM95}{Freedom of Capital Movements 1995} 
}}
\source{
Scholing, E. und Timmermann,  V. (2000): Political and Economic Freedom
}
\usage{
poloecfree
}
\description{
A dataset containing the values  of political an economical situation for 91 countries in 1975 an 1995.
}
\keyword{datasets}
