% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.r
\docType{data}
\name{simplemodel}
\alias{simplemodel}
\title{Simulated data.}
\format{A data frame with 9 variables and 50 cases:
\describe{
  \item{V1,V2,V3}{Indicators of first exogeneous composite} 
  \item{V4,V5,V6}{Indicators of second exogeneous composite}
  \item{V7,V8,V9}{Indicators of endogeneous composite} 
}}
\usage{
simplemodel
}
\description{
The data were simulated with a gsc model with two
exogeneous and one endogeneous compostes. Each composite has three indicators.
All have loadings. There are 50 observations.
}
\keyword{datasets}
